- expanded WRITE_BYTE_TO_SOUND_CHIP macro to allow for muting - removed unneeded convenience variables - expanded songActive to include active and mute bits - reduced curBytes and curFrame back to bytes in TI version to save RAM - did hand edit of CPlayerTI.asm - added several calling examples to tisnplay.c
It's getting close, but it looks like as hacky as it was, the original
player was pretty damn tight. We beat the original compressor for MOST
cases and I'll see if I can get a little closer - we beat the original
on RAM usage and we're close on program size (900 bytes vs 600). CPU
usage is closing in - the worst case seems much better but we're still
above the best and average cases.

Still not sure why 680 rock doesn't compress better though.. the new
vibrato modes should have been right up its alley...