- fix call in getCompressedByteWrap for test code - move type stats out into player and comment out - eliminate DAT0001 using free register R6 - undo the RLEBUF concept - forgot that each stream would need one! - hammer down the highest nails in CPU usage - we now beat the old code on MAX and AVG timing - fix entry to SongLoop - was testing R1 without loading it - make it possible to move song data to the scratchpad - though the CPU gain was only about 5%
This version is working correctly and the CPU usage beats V1 by the
tiniest of margins, but it is faster. (Except for the minimum case -
when nothing is happening it's still 500 cycles slower).

I'm mostly satisfied, but I still need to address the compression
issues. While compression is /usually/ better, when it's worse it's
markedly worse and I need to address that.